Moving execute_and_import_block and ImportResult
